requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django==4.1; python_version >= '3.8'
djangorestframework==3.13.1
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2; python_version >= '3.5'
